Changed builder to restore missing *.py files in Host/pydCaller.
Yuan removed the *.py files in Host/pydCaller to hide more of the
code.  These were replaced by *.pyo files.  The code works but
the ini files all have to be changed to run *.pyo instead of *.py.
I have changed the builder so that both *.py and *.pyo are in
pydCaller so existing config files still work.

+++ bldsup/Builder.py
@@ -172,7 +172,7 @@ class Builder(object):
             fc = os.path.splitext(f)[0] + ".c"
             if os.path.exists(fc):
                 os.remove(fc)
-        self.compile_sources_to_pyo()
+        #self.compile_sources_to_pyo()
 
     def compile_sources_to_pyo(self):
         self.logger.info("Compiling remaining python sources to pyo files")
@@ -180,10 +180,13 @@ class Builder(object):
         self.run_command("python -O -m compileall -x '__init__.py' %s" % path)
         # delete python source files
         for root, dirs, files in os.walk(path, topdown=False):
-            for fname in files:
-                if fname.endswith(".py") and fname != "__init__.py":
-                    fp = os.path.join(root, fname)
-                    os.remove(fp)
+            if "pydCaller" in root:
+                pass
+            else:
+                for fname in files:
+                    if fname.endswith(".py") and fname != "__init__.py":
+                        fp = os.path.join(root, fname)
+                        os.remove(fp)
 
 
     def publish(self):
